fix Battery State for Apple Devices
