+++ app/src/main/java/de/seemoo/at_tracking_detection/database/models/device/BaseDevice.kt
@@ -149,10 +149,9 @@ data class BaseDevice(
             return when (DeviceManager.getDeviceType(scanResult)) {
                 DeviceType.GALAXY_SMART_TAG -> SamsungDevice.getBatteryState(scanResult)
                 DeviceType.GALAXY_SMART_TAG_PLUS -> SamsungDevice.getBatteryState(scanResult)
-                DeviceType.FIND_MY -> AppleDevice.getBatteryState(scanResult)
-                DeviceType.AIRTAG -> AppleDevice.getBatteryState(scanResult)
-                DeviceType.AIRPODS -> AppleDevice.getBatteryState(scanResult)
-                DeviceType.APPLE -> AppleDevice.getBatteryState(scanResult)
+                DeviceType.FIND_MY -> AirTag.getBatteryState(scanResult)
+                DeviceType.AIRTAG -> AirTag.getBatteryState(scanResult)
+                DeviceType.AIRPODS -> AirTag.getBatteryState(scanResult)
                 else -> BatteryState.UNKNOWN
             }
         }
